requirements.txt
# GUI framework dependencies
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio playback
pygame>=2.5.0

# Audio processing
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
moviepy>=1.0.3

# ElevenLabs API integration
elevenlabs>=1.0.0
python-dotenv>=1.0.0

# CLI and utilities
click>=8.1.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
